@@ -25,122 +26,39 @@
         "node": ">=6.0.0"
     },
     "scripts": {
-        "lint": "tslint -t codeFrame 'src/**/*.ts' 'test/**/*.ts'",
-        "prebuild": "rimraf dist",
-        "build": "webpack",
+        "lint": "eslint --ext .ts src/**",
+        "prebuild": "rimraf dist coverage",
+        "build": "parcel build src/data-worker.ts",
         "start": "tsc -w",
-        "test": "jest",
+        "test": "rimraf dist coverage && jest",
         "test:watch": "jest --watch",
         "test:prod": "npm run lint && npm run test -- --coverage --no-cache",
         "docs": "rimraf dist/docs && typedoc --out dist/docs src/data-worker.ts",
         "deploy-docs": "npm run docs && ts-node tools/gh-pages-publish",
-        "report-coverage": "cat ./coverage/lcov.info | coveralls",
-        "commit": "git-cz",
-        "semantic-release": "semantic-release",
-        "semantic-release-prepare": "ts-node tools/semantic-release-prepare",
-        "precommit": "lint-staged",
-        "prepush": "npm run test:prod && npm run build",
-        "commitmsg": "validate-commit-msg"
-    },
-    "lint-staged": {
-        "{src,test}/**/*.ts": [
-            "prettier --write --no-semi --single-quote",
-            "git add"
-        ]
-    },
-    "release": {
-        "analyzeCommits": {
-            "preset": "angular",
-            "releaseRules": [
-                {
-                    "type": "docs",
-                    "scope": "readme",
-                    "release": "patch"
-                },
-                {
-                    "type": "refactor",
-                    "release": "patch"
-                },
-                {
-                    "type": "style",
-                    "release": "patch"
-                }
-            ],
-            "parserOpts": {
-                "noteKeywords": [
-                    "BREAKING CHANGE",
-                    "BREAKING CHANGES",
-                    "BREAKING"
-                ]
-            }
-        }
-    },
-    "config": {
-        "commitizen": {
-            "path": "node_modules/cz-conventional-changelog"
-        },
-        "validate-commit-msg": {
-            "types": "conventional-commit-types",
-            "helpMessage": "Use \"npm run commit\" instead, we use conventional-changelog format :) (https://github.com/commitizen/cz-cli)"
-        }
-    },
-    "jest": {
-        "transform": {
-            ".(ts|tsx)": "<rootDir>/node_modules/ts-jest/preprocessor.js"
-        },
-        "testRegex": "(/__tests__/.*|\\.(test|spec))\\.(ts|tsx|js)$",
-        "moduleFileExtensions": [
-            "ts",
-            "tsx",
-            "js"
-        ],
-        "coveragePathIgnorePatterns": [
-            "/node_modules/",
-            "/test/"
-        ],
-        "coverageThreshold": {
-            "global": {
-                "branches": 90,
-                "functions": 90,
-                "lines": 90,
-                "statements": 90
-            }
-        },
-        "collectCoverage": true,
-        "mapCoverage": true
+        "prepush": "npm run test:prod && npm run build"
     },
     "devDependencies": {
-        "@types/d3-array": "^2.0.0",
+        "@parcel/transformer-typescript-tsc": "^2.3.2",
+        "@types/d3-array": "^3.0.2",
         "@types/d3-collection": "^1.0.8",
-        "@types/jest": "^24.0.15",
-        "@types/node": "^12.6.1",
-        "colors": "^1.3.3",
-        "commitizen": "^3.1.1",
-        "coveralls": "^3.0.4",
-        "cross-env": "^5.2.0",
-        "cz-conventional-changelog": "^2.1.0",
-        "husky": "^3.0.0",
-        "jest": "^24.8.0",
-        "lint-staged": "^9.1.0",
-        "lodash.camelcase": "^4.3.0",
-        "prettier": "^1.18.2",
-        "prompt": "^1.0.0",
-        "replace-in-file": "^4.1.1",
-        "rimraf": "^2.6.3",
-        "semantic-release": "^15.13.18",
-        "ts-jest": "^24.0.2",
-        "ts-loader": "^6.0.4",
-        "ts-node": "^8.3.0",
-        "tslint": "^5.18.0",
-        "typedoc": "^0.14.2",
-        "typescript": "^3.5.3",
-        "uglifyjs-webpack-plugin": "^2.1.3",
-        "validate-commit-msg": "^2.14.0",
-        "webpack": "^4.35.3",
-        "webpack-cli": "^3.3.5"
+        "@types/jest": "^27.4.1",
+        "@types/lodash": "^4.14.181",
+        "@types/node": "^17.0.21",
+        "@typescript-eslint/eslint-plugin": "^5.13.0",
+        "@typescript-eslint/parser": "^5.13.0",
+        "eslint": "^8.10.0",
+        "husky": "^7.0.4",
+        "jest": "^27.5.1",
+        "lodash": "^4.17.21",
+        "parcel": "^2.3.2",
+        "prettier": "^2.5.1",
+        "rimraf": "^3.0.2",
+        "ts-jest": "^27.1.3",
+        "typedoc": "^0.22.13",
+        "typescript": "^4.6.2"
     },
     "dependencies": {
-        "d3-array": "^2.2.0",
+        "d3-array": "^3.0.2",
         "d3-collection": "^1.0.7"
     }
 }
